# Changelog
## Features
- Implemented bookmark synchronization system
  - Added `syncBookmarks` function in `SyncBookmarksUseCase`
  - Updated `FeedViewModel` to handle synchronization
  - Added synchronization response handling in `BookmarksRepositoryImpl`
- Improved local database bookmark management
  - New methods in `BookmarksDao` for updating and retrieving bookmarks
- Updated preferences system to handle last synchronization timestamp

## Improvements
- Refactored `FeedContent` to handle real-time bookmark updates
- Updated Paging Compose version to 3.3.2
- Increased minimum SDK to 26 (Android 8.0)

## API Changes
- Modified `SyncBookmarksResponseDTO` structure to include a `message` object
- Updated synchronization endpoint in `RetrofitNetwork`

## Bug Fixes
- Fixed handling of deleted bookmarks during synchronization

## Testing
- Added new unit tests for `BookmarksDaoTest`
- Updated existing tests to reflect changes in data structure

## Dependencies
- Updated versions of several dependencies, including Paging Compose

## Other
- General refactoring and code cleanup in various files
- Improved documentation in several classes and methods

+++ model/src/main/java/com/desarrollodroide/model/SyncBookmarksRequestPayload.kt
@@ -2,6 +2,6 @@ package com.desarrollodroide.model
 
 data class SyncBookmarksRequestPayload(
     val ids: List<Int>,
-    val last_sync: Int,
+    val last_sync: Long,
     val page: Int
 )
